Remove unnecessary Clients.tsx and display ClientList and ClientId in App.tsx
